corrected the example code in the README to use the actually existing methods cleaned up the public facing API by removing the option to create a builder with invalid state
