+++ MaterialTheming.Tests/ThemeCreation/ThemeCreationService.cs
@@ -15,7 +15,7 @@ internal class ThemeCreationService
     public static Theme CreateTheme(IThemeCreationParameters parameters)
     {
         var mode = parameters.IsDark ? ThemeMode.Dark : ThemeMode.Light;
-        return ThemeBuilder.CreateForColorPaletteFromSourceColor(parameters.SourceColor)
+        return ThemeBuilder.CreateFromSourceColor(parameters.SourceColor)
             .WithMode(mode)
             .WithContrastLevel(parameters.ContrastLevelValue)
             .WithVariant(parameters.Variant)
