+++ MaterialTheming/ThemeBuilder.cs
@@ -1,0 +1,27 @@
+﻿using MaterialTheming.ColorDefinitions;
+using MaterialTheming.Creation;
+
+namespace MaterialTheming;
+
+public class ThemeBuilder
+{
+    /// <inheritdoc cref="ColorPaletteThemeBuilder.CreateFromSourceColor(string)"/>
+    public static IColorPaletteThemeBuilder CreateFromSourceColor(string htmlColor)
+        => ColorPaletteThemeBuilder.CreateFromSourceColor(htmlColor);
+
+    /// <inheritdoc cref="ColorPaletteThemeBuilder.CreateFromSourceColor(HctColor)"/>
+    public static IColorPaletteThemeBuilder CreateFromSourceColor(HctColor color)
+        => ColorPaletteThemeBuilder.CreateFromSourceColor(color);
+
+    /// <inheritdoc cref="ColorPaletteThemeBuilder.CreateFromSourceColor(RgbColor)"/>
+    public static IColorPaletteThemeBuilder CreateFromSourceColor(RgbColor color)
+        => ColorPaletteThemeBuilder.CreateFromSourceColor(color);
+
+    /// <inheritdoc cref="MaterialThemeBuilderJsonThemeBuilder.CreateFromJsonContent(string)"/>
+    public static IMaterialThemeBuilderJsonThemeBuilder CreateFromJsonContent(string materialThemeBuilderJson)
+        => MaterialThemeBuilderJsonThemeBuilder.CreateFromJsonContent(materialThemeBuilderJson);
+
+    /// <inheritdoc cref="MaterialThemeBuilderJsonThemeBuilder.CreateFromJsonFilePath(string)"/>
+    public static IMaterialThemeBuilderJsonThemeBuilder CreateFromJsonFilePath(string materialThemeBuilderJsonFilePath)
+        => MaterialThemeBuilderJsonThemeBuilder.CreateFromJsonFilePath(materialThemeBuilderJsonFilePath);
+}
